requirements.txt
numpy
opencv-python
python-socketio
flask-socketio
eventlet
flask
flask_cors